+++ src/main/java/Work1/Project1/Package/services/EmployeeServices.java
@@ -1,19 +1,31 @@
 package Work1.Project1.Package.services;
 
-import Work1.Project1.Package.entity.DepartmentPK;
-import Work1.Project1.Package.entity.EmployeeEntity;
-import Work1.Project1.Package.entity.EmployeePK;
+import Work1.Project1.Package.converter.EmployeeEntityListToResponseEmpList;
+import Work1.Project1.Package.entity.*;
+import Work1.Project1.Package.repository.CompanyRepository;
+import Work1.Project1.Package.repository.DepartmentRepository;
+import Work1.Project1.Package.requestresponseobject.RequestEmployeeEntity;
 import Work1.Project1.Package.repository.EmployeeRepository;
+import Work1.Project1.Package.requestresponseobject.ResponseDepartmentEntity;
+import Work1.Project1.Package.requestresponseobject.ResponseEmployeeEntity;
+import org.postgresql.util.PSQLException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.UnexpectedRollbackException;
 
 import javax.transaction.Transactional;
+import java.util.HashMap;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import static Work1.Project1.Package.constants.ApplicationConstants.*;
@@ -22,32 +34,86 @@ import static Work1.Project1.Package.constants.ApplicationConstants.*;
 @CacheConfig(cacheNames={"employee_cache"})
 @Component
 @Transactional
-public class EmployeeServices {
+public class EmployeeServices  {
 
     @Autowired
     private EmployeeRepository employeeRepository;
 
+    @Autowired
+    CompanyRepository companyRepository;
 
-    public List<EmployeeEntity> getAllDetails() {
-        return this.employeeRepository.findAll();
+    @Autowired
+    DepartmentRepository departmentRepository;
+    @Autowired
+    EmployeeEntityListToResponseEmpList employeeEntityListToResponseEmpList;
+
+    public Object getAllDetails(Integer pageNo, Integer pageSize) {
+        Pageable paging = PageRequest.of(pageNo, pageSize);/*
+ List<CompanyEntity> allCompanyEntity= companyRepository.findAll(paging);
+ HashMap<Long,HashMap <Long,List<ResponseEmployeeEntity>>> companyToDeptToEmpMapp= new HashMap<Long,HashMap<Long,List<ResponseEmployeeEntity>>>();  //l.getCompanyId()>
+   allCompanyEntity.forEach((c)->{
+                    List<DepartmentEntity> allDepartmentEntities= departmentRepository.findAllByDepartmentPKCompanyId(c.getCompanyId());
+            HashMap <Long,List<ResponseEmployeeEntity>> deptToEmployeemap=new HashMap<Long,List<ResponseEmployeeEntity>>();
+     allDepartmentEntities.forEach((d)-> {
+                        List<EmployeeEntity> employeeEntities = employeeRepository.findByEmployeePKCompanyIdAndEmployeePKDepartmentId(c.getCompanyId(), d.getDepartmentPK().getDepartmentId());
+                        List<ResponseEmployeeEntity> responseEmployeeEntities = employeeEntityListToResponseEmpList.convert(employeeEntities);
+                        deptToEmployeemap.put(d.departmentPK.getDepartmentId(), responseEmployeeEntities);
+                    });
+                    companyToDeptToEmpMapp.put(c.getCompanyId(),deptToEmployeemap);
+                } );
+*/
+        Page<CompanyEntity> allCompanyEntity= companyRepository.findAll(paging);
+        HashMap<Long,HashMap <Long,List<ResponseEmployeeEntity>>> companyToDeptToEmpMapp= new HashMap<Long,HashMap<Long,List<ResponseEmployeeEntity>>>();  //l.getCompanyId()>
+        allCompanyEntity.forEach((c)->{
+            List<DepartmentEntity> allDepartmentEntities= departmentRepository.findAllByDepartmentPKCompanyId(c.getCompanyId());
+            HashMap <Long,List<ResponseEmployeeEntity>> deptToEmployeemap=new HashMap<Long,List<ResponseEmployeeEntity>>();
+            allDepartmentEntities.forEach((d)-> {
+                List<EmployeeEntity> employeeEntities = employeeRepository.findByEmployeePKCompanyIdAndEmployeePKDepartmentId(c.getCompanyId(), d.getDepartmentPK().getDepartmentId());
+                List<ResponseEmployeeEntity> responseEmployeeEntities = employeeEntityListToResponseEmpList.convert(employeeEntities);
+                deptToEmployeemap.put(d.departmentPK.getDepartmentId(), responseEmployeeEntities);
+            });
+            companyToDeptToEmpMapp.put(c.getCompanyId(),deptToEmployeemap);
+        } );
+        if(companyToDeptToEmpMapp.isEmpty())
+            return null;
+        return  companyToDeptToEmpMapp;
     }
 
     @Cacheable(value = "employee_cache")
-    public List<EmployeeEntity> getEmployeeDetails(EmployeePK employeePK) {
-        return this.employeeRepository.findByEmployeePK(employeePK);
-    }
+    public Object getEmployeeDetails(EmployeePK employeePK) {
+        if(employeeRepository.existsById(employeePK)) {
+            Optional<EmployeeEntity> employeeEntity= employeeRepository.findById(employeePK);    //HttpStatus.OK
+            EmployeeEntity employeeEntity1= employeeEntity.get();
+            return( new ResponseEmployeeEntity(employeeEntity1.getEmployeePK().getCompanyId(),employeeEntity1.getEmployeePK().getDepartmentId(),
+                    employeeEntity1.getEmployeePK().getEmployeeId(),employeeEntity1.getEmpName(),employeeEntity1.getPhone(),employeeEntity1.getSalary()) );
+       }
+        else {
+            return null;
+        }
+}
 
-    public String addEmployee(EmployeeEntity employeeEntity) {
+    public String addEmployee(RequestEmployeeEntity requestEmployeeEntity) {
         try {
-            EmployeePK employeePK = employeeEntity.getEmployeePK();
-            if (!employeeRepository.existsById(employeePK)) {
-                this.employeeRepository.save(employeeEntity);
-                return Add_Success;
-            } else {
-                this.updateDetails(employeeEntity);
-                return Update_Success;
+            EmployeePK employeePK =  new EmployeePK(requestEmployeeEntity.getCompanyId(),requestEmployeeEntity.getDepartmentId(),requestEmployeeEntity.getEmployeeId());//employeeEntity.getEmployeePK();
+            EmployeeEntity employeeEntity=new EmployeeEntity(employeePK,requestEmployeeEntity.getEmpName(),requestEmployeeEntity.getPhone(),requestEmployeeEntity.getSalary());
+           // DepartmentPK departmentPK=new DepartmentPK(requestEmployeeEntity.getCompanyId(),)
+           if (!employeeRepository.existsById(employeePK)) {
+                try {
+                    this.employeeRepository.saveAndFlush(employeeEntity);
+                    return Add_Success;
+                }catch(Exception e) {
+                  return Add_Failed;
+                }
+            }else {
+               try {
+                   this.updateDetails(employeeEntity);
+                   return Update_Success;
+               }
+                catch(RuntimeException e) {
+                      return Add_Failed;
+                   }
             }
-        } catch (Exception e) {
+        } catch (RuntimeException e) {
             System.out.println(e);
             return Add_Failed;
         }
@@ -55,13 +121,16 @@ public class EmployeeServices {
 
     @CacheEvict(value = "employee_cache", allEntries = true)
     public String deleteEmployeeDetails(EmployeePK employeePK) throws Exception {
-        try {
-            this.employeeRepository.deleteByEmployeePK(employeePK);
-            return Delete_Success;
-        } catch (Exception e) {
-            // System.out.println("error ms" + e);
-
-            return Delete_Failed;
+        if(employeeRepository.existsById(employeePK)){
+            try {
+                this.employeeRepository.deleteByEmployeePK(employeePK);
+                return Delete_Success;
+            }catch(Exception e)
+            {
+                return Failed;
+            }
+        } else{
+            return Failed;
         }
     }
 
@@ -91,7 +160,7 @@ public class EmployeeServices {
             return Update_Success;
         } catch (Exception e) {
             System.out.println("error ms" + e);
-            return Update_Failed;
+            return Failed;
         }
